Remove repeated code and redundant checks
Taking the answer to lowercase and then comparing saves you the time of checking several combinations.
Also, the current version would not accept answers like 'Tt' or 'sT'.

About menu_description(), the only option where you didn't ran it closed the app, so you are essentially always running it. Moved it outside the conditional.

Also, you wrote a whiteline on every condition  of the main loop, excepting in the about me, so moved the text to the about_me method and put the whiteline only once outside the conditional.